pyproject.toml
[project]
name = "bazarr-auto-translate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
